==> notebooks/03_model+score.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score, roc_curve

DATASET = "../data/train_clean.csv"
COLUMN_TARGET = "SeriousDlqin2yrs"

df = pd.read_csv(DATASET)
df.head()

df.describe()

 # Separar target (y) e variáveis explicativas (X)
 # y = o que queremos prever: 1 = default / 0 = não default
 # Conversão para não prever valores flaot
y = df[COLUMN_TARGET].astype(int)

 # X = todas as colunas menos o target
X = df.drop(columns=[COLUMN_TARGET])
print("Shape X (linhas, colunas):", X.shape)
print("Taxa de default (y.mean):", y.mean())

 # Split treino/teste
 # test_size=0.2: 80% treino, 20% teste
 # stratify=y: garante que a proporção de default no treino e teste fique parecida
 # como apenas 6% é inadimplente, se o stratify=y não for selecionado isso pode acontecer 
X_train, X_test, y_train, y_test = train_test_split(
     X, y, test_size=0.2, random_state=50, stratify=y
)
print("\nTaxa de default treino:", round(y_train.mean(), 3))
print("Taxa de default teste:", round(y_test.mean(), 3))

# Taxa de default treino: 0.067
# Taxa de default teste: 0.067
# Está proporcional!

# Modelo baseline: Regressão Logística
# A regressão logistica é usada quando um resultado é binário
# class_weight='balanced' ajuda o modelo a não ignorar default,
# Com o default é minoritario, ele pode ser ignorado

model = LogisticRegression(max_iter=2000, class_weight="balanced")
model.fit(X_train, y_train)
# Previsão de PD (probabilidade de default)
# predict_proba retorna uma lista com[probabilidade de 0, probabilidade de 1]

pd_test = model.predict_proba(X_test)[:, 1]
# print(f"pd_test={pd_test}")

# AUC (área sob a curva ROC)
# Mede a capacidade do modelo de ranquear bons vs ruins.
# O AUC responde: “Se eu pegar 1 pessoa que deu default e 1 pessoa que não deu, qual a chance do modelo dar um PD maior para a pessoa que deu default?”
# 0.5 = aleatório (ou seja, não sabe nada), 0.7 (Acerta 70 % das vezes), 1.0 = Acerta 100%
# Quanto maior, melhor
auc = roc_auc_score(y_test, pd_test)
print("\nAUC:", auc)
#O Modelo acerta 83% das vezes

# KS (Kolmogorov-Smirnov)
# Métrica muito usada em crédito.
# Responde: “em qual ponto eu separo melhor os ruins dos bons, e quão grande é essa separação?”
# TPR : "Dos que realmente são 1 (defaults), quantos eu capturei como 1"
# FPR : "Dos que realmente são 0 (não-default), quantos eu errei chamando de 1?"
# KS = maior diferença entre TPR e FPR ao longo do ROC.
fpr, tpr, _ = roc_curve(y_test, pd_test)
ks = np.max(tpr - fpr)
print("KS:", ks)


# Tabela por decis (10 grupos por risco)
# Irei ordenar as pessoas pelo risco e divide em 10 grupos (decis)
# Se o modelo estiver bom, o 1º decil (maior PD) deve ter bad_rate bem maior que o último decil (menor PD)
# O primeiro decil (maior PD) deve ter maior bad rate se o modelo estiver separando risco.
# qcut divide a coluna pd em 10 faixas com o mesmo número de pessoas (quantis)
# cada decil tem 10% dos clientes.

tmp = pd.DataFrame({"y": y_test.values, "pd": pd_test})

 # qcut separa em quantis (mesmo número de pessoas por grupo)
tmp["decil"] = pd.qcut(tmp["pd"], 10, duplicates="drop")

# Dicionário
# total: quantas pessoas tem no grupo
# bads: quantos defaults reais aconteceram no grupo (y=1)
# bad_rate: % de default real no grupo (bads/total)
# avg_pd: PD média prevista pelo modelo no grupo (deve ser maior no topo e menor no fim)
decis = tmp.groupby("decil").agg(
    total=("y", "size"),
    bads=("y", "sum"),
    bad_rate=("y", "mean"),
    avg_pd=("pd", "mean"),
).reset_index()

 # Ordena do maior risco para o menor risco
decis = decis.sort_values("avg_pd", ascending=False).reset_index(drop=True)
# Lift vs base: mede “quantas vezes” o decil é mais/menos arriscado que a média.
# base_rate é a taxa média de default na base, lift = bad_rate_do_decil / base_rate (
base_rate = tmp["y"].mean()
decis["lift_vs_base"] = decis["bad_rate"] / (base_rate + 1e-9)
print("\nTabela por decis maior> menor):")
print(decis)
# Interpretação (coeficientes do modelo)
# Na regressão logística:
# coeficiente positivo -> aumenta a PD (mais risco)
# coeficiente negativo -> diminui a PD (menos risco)
coef = pd.Series(model.coef_[0], index=X.columns).sort_values()
print("\nVariáveis que REDUZEM risco (coef negativo):")
print(coef.head(6))
print("\nVariáveis que AUMENTAM risco (coef positivo):")
print(coef.tail(6))


"""
OUTPUT

Shape X (linhas, colunas): (149999, 12)
Taxa de default (y.mean): 0.06684044560297069

Taxa de default treino: 0.067
Taxa de default teste: 0.067

AUC: 0.8393155974147504
KS: 0.5251071285886018

Tabela por decis maior> menor):
            decil  total  bads  bad_rate    avg_pd  lift_vs_base
0     (0.72, 1.0]   3000   995  0.331667  0.835071      4.962593
1   (0.584, 0.72]   3000   351  0.117000  0.653370      1.750623
2  (0.448, 0.584]   3000   233  0.077667  0.513764      1.162095
3  (0.347, 0.448]   3000   143  0.047667  0.393959      0.713217
4  (0.278, 0.347]   3000    97  0.032333  0.309957      0.483791
5  (0.233, 0.278]   3000    55  0.018333  0.254175      0.274314
6  (0.196, 0.233]   3000    57  0.019000  0.214017      0.284289
7  (0.163, 0.196]   3000    30  0.010000  0.178981      0.149626
8   (0.13, 0.163]   3000    26  0.008667  0.146599      0.129676
9  (0.0369, 0.13]   3000    18  0.006000  0.107523      0.089776

Variáveis que REDUZEM risco (coef negativo):
NumberOfTime60-89DaysPastDueNotWorse   -0.299235
Dependents_missing                     -0.055449
age                                    -0.017301
DebtRatio                              -0.000107
MonthlyIncome                          -0.000044
MonthlyIncome_missing                   0.039141
dtype: float64

Variáveis que AUMENTAM risco (coef positivo):
NumberOfOpenCreditLinesAndLoans         0.047236
NumberOfDependents                      0.048321
NumberRealEstateLoansOrLines            0.128114
NumberOfTime30-59DaysPastDueNotWorse    0.291752
NumberOfTimes90DaysLate                 1.135203
RevolvingUtilizationOfUnsecuredLines    2.360617

"""

import numpy as np


# Quanto maior o score, menos risco 
# Multiplicador de score setado para 100

score = np.round((1 - np.array(pd_test)) * 1000).astype(int) # Inversamente proporcial pois o PD é chance de default, o Score alto é comumente visto com algo bom

print("Quantidade de scores:", len(score))
print("Score min/max:", score.min(), score.max())
print("Primeiros 10 scores:", score[:10])

CUTOFF = 500  # nota de corte - aprovado se score >= 500 

approved = (score > CUTOFF).astype(int)
print("\nCutoff:", CUTOFF)
print(f"Aprovados: {round(approved.mean(),2)*100}%")
